+++ packages/next/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @workflow/next
 
+## 4.0.1-beta.9
+
+### Patch Changes
+
+- 4a821fc: Fix Windows path handling by normalizing backslashes to forward slashes in workflow IDs
+- Updated dependencies [4a821fc]
+- Updated dependencies [4a821fc]
+  - @workflow/swc-plugin@4.0.1-beta.3
+  - @workflow/builders@4.0.1-beta.5
+  - @workflow/core@4.0.1-beta.8
+
+## 4.0.1-beta.8
+
+### Patch Changes
+
+- 5dfa4eb: Extract queue trigger configuration constants
+- 2363e47: Move NextBuilder to @workflow/next package
+- Updated dependencies [80d68b7]
+- Updated dependencies [744d82f]
+- Updated dependencies [ebee7f5]
+- Updated dependencies [652485a]
+- Updated dependencies [4585222]
+- Updated dependencies [10bfd4a]
+- Updated dependencies [5dfa4eb]
+- Updated dependencies [05714f7]
+- Updated dependencies [f8c779e]
+- Updated dependencies [bf54a7b]
+  - @workflow/builders@4.0.1-beta.4
+  - @workflow/core@4.0.1-beta.7
+
+## 4.0.1-beta.7
+
+### Patch Changes
+
+- f973954: Update license to Apache 2.0
+- Updated dependencies [10309c3]
+- Updated dependencies [f973954]
+- Updated dependencies [a3326a2]
+  - @workflow/core@4.0.1-beta.6
+  - @workflow/swc-plugin@4.0.1-beta.2
+  - @workflow/cli@4.0.1-beta.7
+
+## 4.0.1-beta.6
+
+### Patch Changes
+
+- Updated dependencies [796fafd]
+- Updated dependencies [70be894]
+- Updated dependencies [20d51f0]
+  - @workflow/core@4.0.1-beta.5
+  - @workflow/cli@4.0.1-beta.6
+
 ## 4.0.1-beta.5
 
 ### Patch Changes